back/pyproject.toml
[tool.poetry]
name = "WebCrawler-back"
version = "0.1.0"
description = ""
authors = ["Thierry Masumbuko"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
alembic = "^1.13.3"
SQLAlchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
uvicorn = "^0.31.0"
pymysql = "^1.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
